+++ packages/atomic-react/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@2.1.0...@coveo/atomic-react@2.1.1) (2023-01-12)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
 # [2.1.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@2.0.12...@coveo/atomic-react@2.1.0) (2023-01-11)
 
 ### Features